Feat: first implementation to solve problem

+++ philo_bonus/src/utils_actions.h
@@ -6,7 +6,7 @@
 /*   By: gvitor-s <gvitor-s>                        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/04/26 12:42:55 by gvitor-s          #+#    #+#             */
-/*   Updated: 2022/04/27 12:52:48 by gvitor-s         ###   ########.fr       */
+/*   Updated: 2022/04/27 18:15:02 by gvitor-s         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,6 +20,6 @@ int				must_die(t_mls last_meal, t_mls die);
 int				ft_atoi(const char *str);
 void			destroy_table(struct s_table **table);
 t_philo			*init_philosopher(int n_seat, char *argv_n_eat);
-struct s_table	*init_table(unsigned long n_philosophers, char *const *argv);
+struct s_table	*init_table(int n_philosophers, char *const *argv);
 void			print_msg(struct s_table *table, char *msg);
 #endif
